feat(profile): split integrations tab into MCP and Claude Chat tabs (PUNT-182)
- Split IntegrationsTab into MCPTab and ClaudeChatTab components
- Rename Integrations tab to MCP, add Claude Chat tab
- Add PathDisplay component with titlebar styling and file icon
- Make CodeBlock copy button always visible (highlights on hover)
- Use relative path in .mcp.json example to match actual config
- Connect PathDisplay and CodeBlock seamlessly with shared borders
