+++ src/app/(app)/profile/profile-content.tsx
@@ -1,12 +1,13 @@
 'use client'
 
-import { KeyRound, Plug, User } from 'lucide-react'
+import { Bot, KeyRound, Terminal, User } from 'lucide-react'
 import Link from 'next/link'
 import { useSearchParams } from 'next/navigation'
 import { useSession } from 'next-auth/react'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { PageHeader } from '@/components/common'
-import { IntegrationsTab } from '@/components/profile/integrations-tab'
+import { ClaudeChatTab } from '@/components/profile/claude-chat-tab'
+import { MCPTab } from '@/components/profile/mcp-tab'
 import { ProfileTab } from '@/components/profile/profile-tab'
 import { SecurityTab } from '@/components/profile/security-tab'
 import { DEMO_USER, isDemoMode } from '@/lib/demo'
@@ -22,9 +23,9 @@ interface UserData {
   isSystemAdmin: boolean
 }
 
-type ProfileTabType = 'profile' | 'security' | 'integrations'
+type ProfileTabType = 'profile' | 'security' | 'mcp' | 'claude-chat'
 
-const VALID_TABS: ProfileTabType[] = ['profile', 'security', 'integrations']
+const VALID_TABS: ProfileTabType[] = ['profile', 'security', 'mcp', 'claude-chat']
 
 function isValidTab(tab: string | null): tab is ProfileTabType {
   return tab !== null && VALID_TABS.includes(tab as ProfileTabType)
@@ -180,16 +181,28 @@ export function ProfileContent() {
             Security
           </Link>
           <Link
-            href="/profile?tab=integrations"
+            href="/profile?tab=mcp"
             className={cn(
               'flex items-center gap-2 px-4 py-2 text-sm font-medium transition-colors border-b-2 -mb-px',
-              activeTab === 'integrations'
+              activeTab === 'mcp'
                 ? 'text-amber-500 border-amber-500'
                 : 'text-zinc-400 border-transparent hover:text-zinc-300',
             )}
           >
-            <Plug className="h-4 w-4" />
-            Integrations
+            <Terminal className="h-4 w-4" />
+            MCP
+          </Link>
+          <Link
+            href="/profile?tab=claude-chat"
+            className={cn(
+              'flex items-center gap-2 px-4 py-2 text-sm font-medium transition-colors border-b-2 -mb-px',
+              activeTab === 'claude-chat'
+                ? 'text-amber-500 border-amber-500'
+                : 'text-zinc-400 border-transparent hover:text-zinc-300',
+            )}
+          >
+            <Bot className="h-4 w-4" />
+            Claude Chat
           </Link>
         </div>
 
@@ -211,7 +224,8 @@ export function ProfileContent() {
               onSessionUpdate={debouncedUpdateSession}
             />
           )}
-          {activeTab === 'integrations' && <IntegrationsTab isDemo={isDemo} />}
+          {activeTab === 'mcp' && <MCPTab isDemo={isDemo} />}
+          {activeTab === 'claude-chat' && <ClaudeChatTab isDemo={isDemo} />}
         </div>
       </div>
